Add files via upload
This project is made using SQL and it is a bit tricky as it has multiple csv files which contains data which are interconnected to each other. And I have written the query by using Join query in it by using multiple csv files. Also there is a word file which contains questionnaire and schema diagram it is like an overview map  of all the csv files.